requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
mcp>=1.0.0
fastmcp>=0.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
redis>=5.0.4
pydantic>=2.5.0
